create new package kyo-test
